+++ R/radae.R
@@ -36,15 +36,15 @@ radae <- function(ADSL, # nolint
                   na_percentage = 0,
                   na_vars = list(AEBODSYS = c(NA, 0.1), AEDECOD = c(1234, 0.1), AETOXGR = c(1234, 0.1)),
                   cached = FALSE) {
-  stopifnot(is.logical.single(cached))
+  stopifnot(is_logical_single(cached))
   if (cached) {
     return(get_cached_data("cadae"))
   }
 
   stopifnot(is.data.frame(ADSL))
-  stopifnot(is.integer.single(max_n_aes))
-  stopifnot(is.null(seed) || is.numeric.single(seed))
-  stopifnot((is.numeric.single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
+  stopifnot(is_integer_single(max_n_aes))
+  stopifnot(is.null(seed) || is_numeric_single(seed))
+  stopifnot((is_numeric_single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
 
   lookup_ae <- if_null( # nolint
     lookup,