+++ R/radrs.R
@@ -37,14 +37,14 @@ radrs <- function(ADSL, # nolint
                   na_percentage = 0,
                   na_vars = list(AVISIT = c(NA, 0.1), AVAL = c(1234, 0.1), AVALC = c(1234, 0.1)),
                   cached = FALSE) {
-  stopifnot(is.logical.single(cached))
+  stopifnot(is_logical_single(cached))
   if (cached) {
     return(get_cached_data("cadrs"))
   }
 
   stopifnot(is.data.frame(ADSL))
-  stopifnot(is.null(seed) || is.numeric.single(seed))
-  stopifnot((is.numeric.single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
+  stopifnot(is.null(seed) || is_numeric_single(seed))
+  stopifnot((is_numeric_single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
 
   param_codes <- if_null(avalc, setNames(1:5, c("CR", "PR", "SD", "PD", "NE")))