+++ R/radsl.R
@@ -44,15 +44,15 @@ radsl <- function(N = 400, # nolint
                     "BMRKR1" = c(seed = 1234, percentage = 0.1), "BMRKR2" = c(1234, 0.1), "BEP01FL" = NA
                   ),
                   cached = FALSE) {
-  stopifnot(is.logical.single(cached))
+  stopifnot(is_logical_single(cached))
   if (cached) {
     return(get_cached_data("cadsl"))
   }
 
-  stopifnot(is.numeric.single(N))
-  stopifnot(is.numeric.single(study_duration))
-  stopifnot(is.null(seed) || is.numeric.single(seed))
-  stopifnot((is.numeric.single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
+  stopifnot(is_numeric_single(N))
+  stopifnot(is_numeric_single(study_duration))
+  stopifnot(is.null(seed) || is_numeric_single(seed))
+  stopifnot((is_numeric_single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
 
   if (!is.null(seed)) {
     set.seed(seed)